# webapp/src/main/java/ar/edu/itba/paw/webapp/controller/resource/roomify-openapi.yaml
openapi: 3.0.3
info:
  title: Roomify API - REST
  description: Official Roomify REST API documentation based on the project specifications.
  version: 1.0.0
servers:
  - url: http://localhost:8080/webapp/api
    description: Local server
...
security:
  - bearerAuth: []
paths:
...
  /contacts:
    get:
      tags: [Contacts]
      summary: List my contacts
      description: Lists contacts or swaps where the authenticated user is involved.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of contacts.
          content:
            application/vnd.roomify.contact.v1.list+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
    post:
      tags: [Contacts]
      summary: Request a swap or booking
      description: Starts a contact request for another user's room.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/vnd.roomify.contact.v1+json:
            schema:
              $ref: '#/components/schemas/ContactCreate'
            example:
              roomRequestedId: 2
              checkIn: '2026-03-03'
              checkOut: '2026-03-06'
              isSwap: true
              roomOfferedId: 3
      responses:
        '201':
          description: Contact created.
          headers:
            Location:
              description: URI of the created contact.
              schema:
                type: string
                format: uri
  /contacts/{id}:
    get:
      tags: [Contacts]
      summary: Get one contact
      description: Returns a single contact where the authenticated user is involved.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ContactId'
      responses:
        '200':
          description: Contact found.
          content:
            application/vnd.roomify.contact.v1+json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          description: Contact not found.
    patch:
      tags: [Contacts]
      summary: Update contact status
      description: Accepts or rejects a pending contact, or cancels a pending/accepted contact. Room swap acceptance requires choosing dates for the offered room.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ContactId'
      requestBody:
        required: true
        content:
          application/vnd.roomify.contact.v1+json:
            schema:
              $ref: '#/components/schemas/ContactUpdate'
            example:
              status: ACCEPTED
              checkIn: '2026-03-03'
              checkOut: '2026-03-06'
      responses:
        '200':
          description: Contact updated.
          content:
            application/vnd.roomify.contact.v1+json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          description: Invalid request body.
        '401':
          description: Not authenticated.
        '403':
          description: Authenticated user cannot update this contact.
        '404':
          description: Contact not found.
        '409':
          description: Contact cannot transition from its current state.
        '422':
          description: Contact update failed domain validation.
...
  /group-trips:
    get:
      tags: [Group Trips]
      summary: List planned trips
      description: Lists the authenticated user's main planned trips.
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema: { type: integer }
        - name: country
          in: query
          schema: { type: string }
        - name: checkIn
          in: query
          schema: { type: string, format: date }
        - name: checkOut
          in: query
          schema: { type: string, format: date }
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 12 }
      responses:
        '200':
          description: List of group trips.
          content:
            application/vnd.roomify.group-trip.v1.list+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupTrip'
    post:
      tags: [Group Trips]
      summary: Create main trip
      description: Starts a new root group trip.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/vnd.roomify.group-trip.v1+json:
            schema:
              $ref: '#/components/schemas/GroupTripCreate'
      responses:
        '201':
          description: Group trip created.
          headers:
            Location:
              description: URI of the created group trip.
              schema:
                type: string
                format: uri
  /group-trips/{id}:
    get:
      tags: [Group Trips]
      summary: Group trip details
      description: Returns the main trip details and includes links to its destinations.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupTripId'
      responses:
        '200':
          description: Group trip details.
          content:
            application/vnd.roomify.group-trip.v1+json:
              schema:
                $ref: '#/components/schemas/GroupTrip'
  /group-trips/{id}/trips:
    post:
      tags: [Group Trips]
      summary: Add destination to trip
      description: Adds a specific destination within a group trip.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupTripId'
      requestBody:
        required: true
        content:
          application/vnd.roomify.group-trip.trip.v1+json:
            schema:
              $ref: '#/components/schemas/TripCreate'
      responses:
        '201':
          description: Destination added.
          headers:
            Location:
              description: URI of the created destination.
              schema:
                type: string
                format: uri
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ContactId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    GroupTripId:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Contact:
      type: object
      properties:
        id: { type: integer }
        roomRequestedId: { type: integer }
        roomOfferedId: { type: integer, nullable: true }
        requestedRange:
          $ref: '#/components/schemas/DateRange'
        offeredRange:
          $ref: '#/components/schemas/DateRange'
        isSwap: { type: boolean }
        moneyOffer: { type: number, format: double }
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED, CANCELED, EXPIRED]
        pendingReview:
          type: boolean
          nullable: true
          description: True when the authenticated user can still leave a review for this past accepted contact.
        self: { type: string, format: uri }
        links:
          type: object
          additionalProperties:
            type: string
            format: uri
          description: Related resource links. Includes review when pendingReview is true.
    ContactCreate:
      type: object
      required: [roomRequestedId, checkIn, checkOut, isSwap]
      properties:
        roomRequestedId: { type: integer }
        checkIn: { type: string, format: date }
        checkOut: { type: string, format: date }
        isSwap: { type: boolean }
        dayPrice: { type: number, format: double }
        roomOfferedId: { type: integer, nullable: true }
    ContactUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ACCEPTED, REJECTED, CANCELED]
        checkIn: { type: string, format: date }
        checkOut: { type: string, format: date }
    DateRange:
      type: object
      properties:
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
...
    GroupTrip:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        tripId: { type: integer, nullable: true }
        country: { type: string, nullable: true }
        countryCode: { type: string, nullable: true }
        tripStartDate: { type: string, format: date, nullable: true }
        tripEndDate: { type: string, format: date, nullable: true }
        trips:
          type: array
          items:
            $ref: '#/components/schemas/Trip'
        links:
          type: array
          items:
            type: object
            properties:
              rel: { type: string }
              href: { type: string }
    GroupTripCreate:
      type: object
      required: [ownerId, title]
      properties:
        ownerId: { type: integer }
        title: { type: string }
    Trip:
      type: object
      properties:
        id: { type: integer }
        destination: { type: string }
        checkIn: { type: string, format: date }
        checkOut: { type: string, format: date }
    TripCreate:
      type: object
      required: [destination, checkIn, checkOut]
      properties:
        destination: { type: string }
        checkIn: { type: string, format: date }
        checkOut: { type: string, format: date }